api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gtm-copilot-api"
version = "0.1.0"
description = "Open-source GTM copilot backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.18",
  "pymysql>=1.1.1",
  "pgvector>=0.3.2",
  "alembic>=1.13.2",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.3.4",
  "python-dotenv>=1.0.1",
  "typer>=0.12.3",
  "celery>=5.4.0",
  "redis>=5.0.7",
  "httpx>=0.27.0",
  "openai>=1.40.0",
  "anthropic>=0.40.0",
  "python-dateutil>=2.9.0",
  "email-validator>=2.2.0",
  "jinja2>=3.1.4"
  ,
  "google-api-python-client>=2.151.0",
  "google-auth>=2.34.0",
  "google-auth-oauthlib>=1.2.1",
  "google-auth-httplib2>=0.2.0",
  "cryptography>=43.0.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib>=1.7.4",
  "tidb-vector>=0.0.9",
  "firecrawl-py>=1.0.0",
  "tiktoken>=0.7.0",
  "slack_sdk>=3.27.0",
  "sentry-sdk[fastapi]==2.22.0",
  "python-json-logger==2.0.7"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "pytest-cov>=5.0.0",
  "anyio>=4.4.0"
]

[project.scripts]
kb = "app.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
